+++ lib/core/message/scheduled_user_message.g.dart
@@ -49,6 +49,7 @@ ScheduledUserMessage _$ScheduledUserMessageFromJson(
       data: json['data'] as String?,
       sendingStatus: $enumDecodeNullable(
           _$MessageSendingStatusEnumMap, json['sending_status']),
+      isPinnedMessage: json['is_pinned_message'] as bool? ?? false,
     )
       ..sender = json['user'] == null
           ? null
@@ -72,6 +73,7 @@ ScheduledUserMessage _$ScheduledUserMessageFromJson(
 Map<String, dynamic> _$ScheduledUserMessageToJson(
         ScheduledUserMessage instance) =>
     <String, dynamic>{
+      'is_pinned_message': instance.isPinnedMessage,
       'message': instance.message,
       'sending_status': _$MessageSendingStatusEnumMap[instance.sendingStatus],
       'user': instance.sender?.toJson(),